pyproject.toml
[project]
name = "eventhorizon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0",
    "django-allauth[socialaccount]>=65.13.1",
    "django-rest-framework>=0.1.0",
    "python-dotenv>=1.2.1",
    "tzdata>=2025.2",
    "django-oauth-toolkit",
    "djangorestframework",
    "markdown",
    "django-filter",
    "django-rest-knox>=5.0.2",
    "django-cors-headers>=4.9.0",
    "pillow>=12.0.0",
]
